// server/src/common/types/file-types.ts
// ^ типы/пути файлов

export enum FileType {
  ALL = 'all',
  IMAGE = 'images',
  ALBUM = 'images/album',
  PICTURE = 'images/picture',
  DRAWING = 'images/drawing',
  USER = 'users',
  AVATAR = 'users/avatar',
  PHOTO = 'users/photo',
  PERSONAL = 'users/personal',
  AUDIO = 'audios',
  TRACK = 'audios/track',
  AUDIOBOOK = 'audios/audiobook',
  SOUND = 'audios/sound',
  FILE = 'files',
  TEXT = 'files/document',
  VIDEO = 'videos',
  PRIVATE = 'videos/private',
  DOCUMENT = 'files/document',
  BOOK = 'files/book',
  PROSE = 'files/prose',
  CODE = 'codes',
  WEB = 'codes/web',
  FULLSTACK = 'codes/fullstack',
  SCHEME = 'codes/scheme',
  BLUEPRINT = 'codes/blueprint',
  APPLICATION = 'application',
  TRASH = 'trash',
  OTHER = 'other',
}